requirements.txt
fastapi
google-generativeai
httpie
jupyter
pandas
pydantic
statsbombpy
streamlit
uvicorn